Add `onTouch` callback to Slider component
This commit introduces an `onTouch` callback to the `Slider` composable.
This allows consumers to be notified when the user interacts with either the slider thumb or track by pressing down on them.

Additionally, this change fixes a minor formatting issue in the bottom text of the slider where an extra space was present before the suffix.